gcp/cloud-functions/pubmed-ingestion/requirements.txt
google-cloud-bigquery==3.14.1
google-cloud-storage==2.14.0
google-cloud-aiplatform==1.40.0
requests==2.31.0
